Create an abstract BasicOutputExecutionResult class
- Make FinishedExecutionResult extend this class
- Include Lombok stuff
