+++ executor-service-core/executor-service-models/src/main/java/ar/edu/itba/cep/executor_service/models/FinishedExecutionResult.java
@@ -1,31 +1,11 @@
 package ar.edu.itba.cep.executor_service.models;
 
-import org.springframework.util.Assert;
-
 import java.util.List;
-import java.util.Objects;
 
 /**
  * Represents an {@link ExecutionResult} that has finished (i.e has exit code, stdout and stderr).
  */
-public class FinishedExecutionResult extends ExecutionResult {
-
-    /**
-     * The execution's exit code.
-     */
-    private final int exitCode;
-
-    /**
-     * A {@link List} of {@link String}s that were sent to standard output by the program being executed.
-     * Each {@link String} in the {@link List} is a line that was printed in standard output.
-     */
-    private final List<String> stdout;
-
-    /**
-     * A {@link List} of {@link String}s that were sent to standard error output by the program being executed.
-     * Each {@link String} in the {@link List} is a line that was printed in standard error output.
-     */
-    private final List<String> stderr;
+public class FinishedExecutionResult extends BasicOutputExecutionResult {
 
 
     /**
@@ -47,76 +27,6 @@ public class FinishedExecutionResult extends ExecutionResult {
             final List<String> stdout,
             final List<String> stderr,
             final ExecutionRequest executionRequest) throws IllegalArgumentException {
-        super(executionRequest);
-        assertStdOutList(stdout);
-        assertStdErrList(stderr);
-        this.exitCode = exitCode;
-        this.stdout = stdout;
-        this.stderr = stderr;
-    }
-
-
-    /**
-     * @return The execution's exit code.
-     */
-    public int getExitCode() {
-        return exitCode;
-    }
-
-    /**
-     * @return A {@link List} of {@link String}s that were sent to standard output by the program being executed.
-     * Each {@link String} in the {@link List} is a line that was printed in standard output.
-     */
-    public List<String> getStdout() {
-        return stdout;
-    }
-
-    /**
-     * @return A {@link List} of {@link String}s that were sent to standard error output by the program being executed.
-     * Each {@link String} in the {@link List} is a line that was printed in standard error output.
-     */
-    public List<String> getStderr() {
-        return stderr;
-    }
-
-
-    // ================================
-    // toString
-    // ================================
-
-    @Override
-    public String toString() {
-        return "FinishedExecutionResult{" +
-                "exitCode=" + exitCode +
-                ", stdout=" + stdout +
-                ", stderr=" + stderr +
-                "} " + super.toString();
-    }
-
-
-    // ================================
-    // Assertions
-    // ================================
-
-    /**
-     * Asserts that the given {@code stdout} {@link List} is valid.
-     *
-     * @param stdout The {@link List} with standard output to be validated.
-     * @throws IllegalArgumentException If the {@link List} is not valid.
-     */
-    private static void assertStdOutList(final List<String> stdout) throws IllegalArgumentException {
-        Assert.notNull(stdout, "The stdout list must not be null");
-        Assert.isTrue(stdout.stream().noneMatch(Objects::isNull), "The stdout list must not contain nulls.");
-    }
-
-    /**
-     * Asserts that the given {@code stderr} {@link List} is valid.
-     *
-     * @param stderr The {@link List} with standard error output to be validated.
-     * @throws IllegalArgumentException If the {@link List} is not valid.
-     */
-    private static void assertStdErrList(final List<String> stderr) throws IllegalArgumentException {
-        Assert.notNull(stderr, "The stderr list must not be null");
-        Assert.isTrue(stderr.stream().noneMatch(Objects::isNull), "The stderr list must not contain nulls.");
+        super(exitCode, stdout, stderr, executionRequest);
     }
 }
